protokoly zrobione

+++ src/main/java/edu/uph/ii/platformy/models/ProtocolItem.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class ProtocolItem
 {
-        private Integer protocolId;
+        private Long protocolId;
         private String firstTermin;
         private String secondTermin;
 }
